[TrueCloudLab#XX] Add billing metrics to separate registry
